+++ packages/crypto/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./crypto";
+export * from "./turnkey";
